rubocops: rename files for consistency.
Make the filenames consistent with the class names.
