+++ Library/Homebrew/rubocops.rb
@@ -1,17 +1,17 @@
 require_relative "load_path"
 
 require "rubocop-rspec"
-require "rubocops/formula_desc_cop"
-require "rubocops/components_order_cop"
-require "rubocops/components_redundancy_cop"
-require "rubocops/dependency_order_cop"
-require "rubocops/homepage_cop"
-require "rubocops/text_cop"
-require "rubocops/caveats_cop"
-require "rubocops/checksum_cop"
-require "rubocops/patches_cop"
-require "rubocops/conflicts_cop"
-require "rubocops/options_cop"
-require "rubocops/urls_cop"
-require "rubocops/lines_cop"
-require "rubocops/class_cop"
+require "rubocops/formula_desc"
+require "rubocops/components_order"
+require "rubocops/components_redundancy"
+require "rubocops/dependency_order"
+require "rubocops/homepage"
+require "rubocops/text"
+require "rubocops/caveats"
+require "rubocops/checksum"
+require "rubocops/patches"
+require "rubocops/conflicts"
+require "rubocops/options"
+require "rubocops/urls"
+require "rubocops/lines"
+require "rubocops/class"
