+++ Library/Homebrew/rubocops/conflicts.rb
@@ -1,4 +1,4 @@
-require "rubocops/extend/formula_cop"
+require "rubocops/extend/formula"
 require "extend/string"
 
 module RuboCop
