+++ Library/Homebrew/rubocops/extend/formula.rb
