+++ Library/Homebrew/test/rubocops/homepage_spec.rb
@@ -1,4 +1,4 @@
-require "rubocops/homepage_cop"
+require "rubocops/homepage"
 
 describe RuboCop::Cop::FormulaAudit::Homepage do
   subject(:cop) { described_class.new }
