+++ Library/Homebrew/test/rubocops/patches_spec.rb
@@ -1,4 +1,4 @@
-require "rubocops/patches_cop"
+require "rubocops/patches"
 
 describe RuboCop::Cop::FormulaAudit::Patches do
   subject(:cop) { described_class.new }
