+++ Library/Homebrew/test/rubocops/text_spec.rb
@@ -1,4 +1,4 @@
-require "rubocops/text_cop"
+require "rubocops/text"
 
 describe RuboCop::Cop::FormulaAudit::Text do
   subject(:cop) { described_class.new }
